+++ front/app/containers/Admin/projects/all/_shared/messages.ts
@@ -1,0 +1,20 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  overviewPageTitle: {
+    id: 'app.containers.AdminPage.ProjectDashboard.overviewPageTitle',
+    defaultMessage: 'Projects',
+  },
+  onlyAdminsCanCreateFolders: {
+    id: 'app.containers.AdminPage.projects.all.onlyAdminsCanCreateFolders',
+    defaultMessage: 'Only admins can create project folders.',
+  },
+  createProjectFolder: {
+    id: 'app.containers.AdminPage.projects.all.createProjectFolder',
+    defaultMessage: 'New folder',
+  },
+  newProject: {
+    id: 'app.containers.AdminPage.ProjectEdit.createProject',
+    defaultMessage: 'New project',
+  },
+});
